[TM-7] - (5/5) - delete dev object by primaryKey

+++ src/main/java/com/tma/teamhr/repository/IDevRepository.java
@@ -3,5 +3,5 @@ package com.tma.teamhr.repository;
 import com.tma.teamhr.model.Dev;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IDevRepository extends CrudRepository<Dev, Long> {
+public interface IDevRepository extends CrudRepository<Dev, Integer> {
 }
